Add files via upload
Creating Cohorts of Songs
Description:
The customer always looks forward to specialized treatment, whether shopping on an e-commerce website or watching Netflix. They want what they might like to see. To keep the customers engaged, it is also crucial for companies to always present the most relevant information. Spotify is a Swedish audio streaming and media service provider. The company has over 456 million active monthly users, including over 195 million paying subscribers, as of September 2022. The company intends to create cohorts of different songs that will aid in the recommendation of songs to users based on various relevant features. Each cohort would contain similar types of songs. This dataset contains data from Spotify's API about all albums for the Rolling Stones listed on Spotify.
Objective:
To perform Exploratory Data Analysis and Cluster Analysis to create cohort of songs. The goal is to understand the various factors that contribute to creating cohort of songs.
Tasks to be performed:
Inspect the dataset and check for duplicate and missing values.
Check for outliers if any and remove them.
Once the data is cleaned and ready, perform EDA to find patterns in the data.
 Perform cluster analysis with K-Means Clustering.
Apply principal Component Analysis for dimensionality reduction.
Analysis:
The dataset did not have any missing values and there were a few duplicate values which were taken care of. The features that did not add any value to the analysis were dropped. The outliers were detected using Box and whiskers plot and removed using IQR method.
After performing EDA, it was found that, ‘Some Girls’ and ‘Sticky Fingers (Remastered) were the most popular albums.
From the heatmap of correlation analysis, it was found that the feature, ’popularity’ was negatively correlated with ‘liveness’, ‘speechiness’, ‘tempo’ and ‘energy’. 
The data needed to be scaled since there was a lot of variation between the features. They were scaled using MInMax Scaler.
Using the ‘Elbow Method’, the number of clusters was found out which stood at 5.
After performing cluster analysis using K-Means clustering, the model had performed poorly because of the large number of features. The Evaluation Metrics of clustering, Silhouette Score showed 47%.
There was a need for dimensionality reduction for the clustering to be satisfying. Again, the number of components of PCA was established based on the cumulative variance plot which was 3.
When K-Means clustering was applied to the reduced data, the Silhouette Score gave 72% goodness. This analysis shows that when we apply dimensionality reduction to a large dataset, then clustering algorithm performs much better than without PCA.