Add files via upload
Capstone Project: Healthcare - Prediction of Diabetes
Description:
NIDDK (National Institute of Diabetes, Digestive and Kidney Diseases) research creates
knowledge about and treatments for the most chronic, costly and consequential diseases.
The dataset used in this project is originally from NIDDK. This dataset consists of information
about female patients who may or may not have got Diabetes during their pregnancies.
Objective:
The objective of this project is to accurately predict whether a patient has diabetes or not
based on certain diagnostic measurements included in the dataset, using a suitable model.
Data Exploration:
On inspecting the dataset, we come to know that,
● There are eight predictor variables such as number of pregnancies a patient has had, BMI,
BP, Insulin levels etc. and one target variable (Outcome).
● Plot histograms of all the variables to visualize the data.
● Some of the variables like, Glucose, BMI, Skin Thickness, Insulin and Blood Pressure
have 0’s in their columns. It is not practically possible to have 0 values. We have to treat
them as missing values and fill them with the mean value of the rest of the values in the
respective columns.
● Create various count plots to understand the data types in the data, count of outcome and
relationship with all the variables with the ‘Outcome’.
● Create pair plots to understand the relationship between the variables.
● Performing Correlation analysis and visualizing with a heatmap.
Data Modeling:
● For training a model, we need to select the features that are correlated. we shall find
features that have correlation greater than or equal to 0.5 and select those features and
train the model. There are five variables that have correlation with the outcome. They are,
Pregnancies, Glucose, BMI, Skin Thickness and Age.
● Using these five variables we shall build a model for predicting the outcome.
● But before that we need to standardize the data. We shall do it using Standard Scaler.
● Three models are trained namely, Logistic Regression model, Decision Tree Classifier
and Random Forest Classifier with a training data to testing data ratio of 80:20.
● On comparing the performance of these models with that of K-NN algorithm, we can
infer that the Logistic Regression model and the K-NN algorithm have the same f1-score of
0.63. Whereas, the Random Forest classifier is slightly better with an f1-score of 0.64.
Decision tree classifier has performed poorly with an f1-score of 0.59.
● Create an AUC (ROC curve) to check how accurately the model is able to predict the
outcome. Ideally, the ROC (Receiver Operating Characteristics) curve should be closer to
the top left point. ROC and AUC or the Area under the ROC curve is used to check a
Classification model’s performance. AUC ranges from 0 to 1, where 1 indicates that the
model is excellent and 0 indicates worst. From the AUC ROC curve, Sensitivity,
Specificity and Accuracy metrics, we can conclude that the Logistic Regression model is
a better Choice for Predicting Diabetes.
Data Visualization:
A Dashboard to analyze the data has been created using Tableau.