Add files via upload
Project on Health Care
Prediction of Cardiovascular Disease using Machine Learning
Cardiovascular disease is a leading cause of death globally. An effective model is required to
predict the heart disease using various parameters that are available for reference that may
have an impact on the heart health.
After doing the preliminary analysis of the dataset, like checking the shape, datatypes and
unique values, it can be concluded that the data is clean without any duplicate and missing
values.
Visualizing the data:
By using visualization tools like Histogram, we can find the Numerical data columns and
Categorical data columns and also the relationship of each feature with the target column.
We can see that 'sex', 'cp', 'fbs', 'restecg', 'exang', 'slope', 'ca', 'thal' and 'target' columns are
categorical columns.
Visualization of Categorical Columns using Count plot:
To check the relationship between the categorical features and the Cardiovascular disease,
we can use Countplot from the seaborn library.
We can conclude from the Count Plots as follows:
● Of the total number of patients, the number of Male patients is higher, but lesser
Male patients with Cardiovascular disease. The count of Female patients is lesser as
compared to Male patients, but more women have Cardiovascular disease.
● The Chest pain type 2 is connected to the Heart disease and most patients with
Chest pain type 0 do not have a Heart disease.
● Patients with fasting blood sugar level < 120, have a Heart disease.
● Most patients with Resting Electrocardiograph of type 1 have a heart disease
followed by type 0.
● Exercise induced Angina = No, have a heart disease.
● Slope of the peak exercise ST segment of type 2 have maximum Heart Disease.
● Number of major vessels (0-3) colored by fluoroscopy of Type 0 has maximum
number of Heart Attacks.
● When we compare CVD with age, we find that CardioVascular disease is highest at
54 years of age, followed by 41, 51 and 52 years.
● Resting Blood Pressure (trestbps): Maximum heart attacks have occured at Resting
Blood Pressure = 120 and 130.
● At cholesterol levels between 195 and 240, there is a good chance of having a heart
attack.
● Peak exercising '2' has a significant impact on Heart Attack.
● Thalassemia of type '2' has a significant impact on the occurance of a Heart Attack.
We check the correlation between all the features using a Heatmap. Very dark and very light
areas mean that there is a correlation between the corresponding variables.
Model Building:
We use One hot encoding to convert categorical data into binary vectors. After splitting the
data into training and testing in the ratio of 80:20, we fit the training data to Logistic
Regression and Random Forest Classifier models, we predict the outcome and check the
accuracy of the models.
Result:
Accuracy of the Random Forest Classifier is slightly better at 78.6885% as compared to that
of Logistic Regression at 77.0492%.
Statistical Analysis:
By using the Logit Regression model from the statts.api, we can do find p-value, standard
error, etc. for all the features and use the result for feature selection. Usually,
features with p-value < 0.05 is considered and also standard error < 1.
We again train the Models using the features selected after statistical analysis.
Result:
The accuracy of the Logistic Regression Model has improved from 77.0492% to 81.9672%.
Whereas, the accuracy of Random Forest Classifier Model has reduced from 78.6885% to
75.4098%.
Feature Importance using Random Forest Classifier:
By using Random Forest Classifier, we check for the important features in the descending
order and select the first five features for training the models.
Result:
The accuracy of Logistic Regression has slightly improved to 83.6066% and the accuracy of
Random Forest Classifier to 81.9672%.