Add files via upload
This project consists of basic data cleaning and Exploratory Data Analysis to find patterns in the complaint types. 
Testing of hypothesis with Chi-square test was done on categorical data to check whether the average response time across all complaint types is similar. 
One sample T-test to check if the type of complaint is related to the location.